ORS/requirements.txt
Django==1.10.1
django-cors-headers==1.1.0
djangorestframework==3.4.7
enum34==1.1.6
future==0.15.2
futures==3.0.5
gax-google-logging-v2==0.8.3
gax-google-pubsub-v1==0.8.3
gcloud==0.18.1
google-gax==0.14.1
googleapis-common-protos==1.3.4
grpc-google-logging-v2==0.9.3
grpc-google-pubsub-v1==0.9.3
grpcio==1.0.0
httplib2==0.9.2
MySQL-python==1.2.5
oauth2client==3.0.0
pip==8.1.2
pkg-resources==0.0.0
ply==3.8
protobuf==3.0.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
rsa==3.4.2
setuptools==27.3.0
six==1.10.0
wheel==0.30.0a0
